requirements.txt
openai==1.54.0
python-dotenv==1.0.1
PyQt5==5.15.11
PyYAML==6.0.2
pyttsx3==2.98
pyinstaller==6.11.1
